requirements.txt
flask>=2.2,<2.3
flask-sqlalchemy>=3.0.2,<3.1
flask-cors>=3.0.10,<3.1
pymysql>=1.0.2,<1.1
python-dotenv>=0.19.2,<0.20
requests>=2.27.1,<2.28
beautifulsoup4>=4.10.0,<4.11
pytz>=2021.3,<2022.0
mysql-connector-python>=8.0.28,<8.1
schedule>=1.1.0,<1.2
